@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
-// 	protoc        v3.6.1
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.27.1
 // source: shipping/shipping.proto
 
 package shipping
@@ -119,10 +119,10 @@ var file_shipping_shipping_proto_rawDesc = []byte{
 	0x72, 0x65, 0x61, 0x74, 0x65, 0x53, 0x68, 0x69, 0x70, 0x70, 0x69, 0x6e, 0x67, 0x52, 0x65, 0x71,
 	0x75, 0x65, 0x73, 0x74, 0x1a, 0x17, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x53, 0x68, 0x69,
 	0x70, 0x70, 0x69, 0x6e, 0x67, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42,
-	0x36, 0x5a, 0x34, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x68, 0x75,
-	0x73, 0x65, 0x79, 0x69, 0x6e, 0x62, 0x61, 0x62, 0x61, 0x6c, 0x2f, 0x6d, 0x69, 0x63, 0x72, 0x6f,
-	0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x73, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x73,
-	0x68, 0x69, 0x70, 0x70, 0x69, 0x6e, 0x67, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x31, 0x5a, 0x2f, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x7a, 0x6c,
+	0x77, 0x38, 0x38, 0x34, 0x34, 0x2f, 0x6d, 0x69, 0x63, 0x72, 0x6f, 0x73, 0x65, 0x72, 0x76, 0x69,
+	0x63, 0x65, 0x73, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x73, 0x68, 0x69, 0x70, 0x70, 0x69,
+	0x6e, 0x67, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -138,7 +138,7 @@ func file_shipping_shipping_proto_rawDescGZIP() []byte {
 }
 
 var file_shipping_shipping_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_shipping_shipping_proto_goTypes = []interface{}{
+var file_shipping_shipping_proto_goTypes = []any{
 	(*CreateShippingRequest)(nil),  // 0: CreateShippingRequest
 	(*CreateShippingResponse)(nil), // 1: CreateShippingResponse
 }
@@ -158,7 +158,7 @@ func file_shipping_shipping_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_shipping_shipping_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_shipping_shipping_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateShippingRequest); i {
 			case 0:
 				return &v.state
@@ -170,7 +170,7 @@ func file_shipping_shipping_proto_init() {
 				return nil
 			}
 		}
-		file_shipping_shipping_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_shipping_shipping_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateShippingResponse); i {
 			case 0:
 				return &v.state
