@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.6.1
+// - protoc-gen-go-grpc v1.4.0
+// - protoc             v5.27.1
 // source: shipping/shipping.proto
 
 package shipping
@@ -15,8 +15,12 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.62.0 or later.
+const _ = grpc.SupportPackageIsVersion8
+
+const (
+	Shipping_Create_FullMethodName = "/Shipping/Create"
+)
 
 // ShippingClient is the client API for Shipping service.
 //
@@ -34,8 +38,9 @@ func NewShippingClient(cc grpc.ClientConnInterface) ShippingClient {
 }
 
 func (c *shippingClient) Create(ctx context.Context, in *CreateShippingRequest, opts ...grpc.CallOption) (*CreateShippingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateShippingResponse)
-	err := c.cc.Invoke(ctx, "/Shipping/Create", in, out, opts...)
+	err := c.cc.Invoke(ctx, Shipping_Create_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -80,7 +85,7 @@ func _Shipping_Create_Handler(srv interface{}, ctx context.Context, dec func(int
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/Shipping/Create",
+		FullMethod: Shipping_Create_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ShippingServer).Create(ctx, req.(*CreateShippingRequest))
